requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
cffi==1.14.4
chardet==4.0.0
idna==2.10
multidict==5.1.0
enhanced-dpy==1.6.0.6
pycparser==2.20
PyNaCl==1.4.0
six==1.15.0
yarl==1.6.3

pyrebase5~=5.0.1
python-dotenv~=0.15.0
APScheduler~=3.6.3
PyGithub~=1.54
ossapi~=1.3.0
python-dateutil~=2.8.1
bs4~=0.0.1
beautifulsoup4~=4.9.3
pytz~=2020.4
pygelbooru~=0.3.1
requests~=2.25.1
pyttanko~=1.2.1